Use the same parameter type with malloc.
Using different types here may lead to various problems during type
conversion.


git-svn-id: svn://git.mplayerhq.hu/mplayer/trunk@35439 b3059339-0415-0410-9bf9-f77b7e298cf2

+++ libmpcodecs/vf_remove_logo.c
@@ -170,7 +170,7 @@ struct vf_priv_s
  * of how MPlayer works, it cannot safely halt execution, but at least the user
  * will get an error message before the segfault happens.
  */
-static void * safe_malloc(int size)
+static void * safe_malloc(size_t size)
 {
   void * answer = malloc(size);
   if (answer == NULL)
